Move AddNewFunctionSubGraph to FlowGraphEditorWindow
