+++ Runtime/Plugins/IL2CPPCall.dll.meta
